service/App/src/requirements.txt
Werkzeug==1.0.1
bleach==4.1.0
Flask==1.1.2
mysql-connector-python==8.0.22
gunicorn==20.1.0
eventlet==0.30.2
